added a root directory file
This file finds the root path for our project. It does this by finding the path of the file itself then moving up two directories.

This should mean that avoid a lot of problems with paths as we try to run our files across multiple machines.

Not the best explanation but feel free to message me if you want me to go into more detail